+++ services/web/server/src/simcore_service_webserver/functions/_services_metadata/proxy.py
@@ -1,5 +1,5 @@
 from ._models import ServiceMetadata
-from ._service import batch_get_service_metadata, get_service_metadata
+from ._proxy import batch_get_service_metadata, get_service_metadata
 
 __all__: tuple[str, ...] = (
     "ServiceMetadata",
